requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiohttp-swagger==1.0.16
aiosignal==1.2.0
apispec==5.1.1
astor==0.8.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bandit==1.7.4
charset-normalizer==2.0.12
contextvars==2.4
coverage==6.4.1
darglint==1.8.1
dataclasses==0.6
docutils==0.18.1
eradicate==2.1.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.4.0
flake8-bugbear==22.4.25
flake8-commas==2.1.0
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.2.1
flake8-isort==4.1.1
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.6
flake8-string-format==0.3.0
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
immutables==0.18
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
mccabe==0.6.1
multidict==6.0.2
opentracing==2.4.0
packaging==21.3
pbr==5.9.0
pep8-naming==0.12.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.18.3
pytest-cov==3.0.0
python-dotenv==0.20.0
PyYAML==5.4.1
restructuredtext-lint==1.4.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==3.5.0
testfixtures==6.18.5
tomli==2.0.1
trafaret==2.1.1
typing_extensions==4.2.0
webargs==8.1.0
wemake-python-styleguide==0.16.1
yarl==1.7.2
